[IMP] Change location view
