+++ pabi_account_report/reports/__init__.py
@@ -1,8 +1,4 @@
 # -*- coding: utf-8 -*
-from . import res_partner_view
-from . import payable_detail_view
-from . import sla_purchase_view
-from . import sla_employee_view
 from . import xlsx_report_cheque_register
 from . import xlsx_report_partner_list
 from . import xlsx_report_partner_detail
